requirements.txt
numpy==1.24.3
Flask==2.2.2
beautifulsoup4==4.12.2
requests==2.31.0
PyPDF2==3.0.1
resend==0.5.2
boto3==1.28.51
python-dotenv==1.0.0
aiohttp==3.8.5